% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowWAFOMNX.R
\name{lowWAFOMNX.dimF2MinMax}
\alias{lowWAFOMNX.dimF2MinMax}
\title{get minimum and maximum F2 dimension number.}
\usage{
lowWAFOMNX.dimF2MinMax(dimR)
}
\arguments{
\item{dimR}{dimention.}
}
\value{
supported minimum and maximum F2 dimension number
}
\description{
get minimum and maximum F2 dimension number.
}
