% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowWAFOMNX.R
\name{lowWAFOMNX.dimMinMax}
\alias{lowWAFOMNX.dimMinMax}
\title{get minimum and maximum dimension number of Low WAFOM Niederreiter-Xing
Sequence}
\usage{
lowWAFOMNX.dimMinMax()
}
\value{
supported minimum and maximum dimension number.
}
\description{
get minimum and maximum dimension number of Low WAFOM Niederreiter-Xing
Sequence
}
