% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowWAFOMNX.R
\name{lowWAFOMNX.points}
\alias{lowWAFOMNX.points}
\title{get points from Low WAFOM Niederreiter-XingSobolSequence}
\usage{
lowWAFOMNX.points(dimR, dimF2 = 10, digitalShift = FALSE)
}
\arguments{
\item{dimR}{dimension.}

\item{dimF2}{F2-dimension of each element.}

\item{digitalShift}{use digital shift or not.}
}
\value{
matrix of points where every row contains dimR dimensional point.
}
\description{
This R version does not returns coordinate value zero,
but returns value very near to zero, 2^-64.
}
